Rename TM trafficopsdata to todata
