+++ traffic_monitor_golang/traffic_monitor/todata/todata.go
@@ -1,4 +1,4 @@
-package trafficopsdata
+package todata
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
@@ -116,8 +116,8 @@ func (d TODataThreadsafe) set(newTOData TOData) {
 type CRConfig struct {
 	ContentServers map[tc.CacheName]struct {
 		DeliveryServices map[tc.DeliveryServiceName][]string `json:"deliveryServices"`
-		CacheGroup       string                                `json:"cacheGroup"`
-		Type             string                                `json:"type"`
+		CacheGroup       string                              `json:"cacheGroup"`
+		Type             string                              `json:"type"`
 	} `json:"contentServers"`
 	DeliveryServices map[tc.DeliveryServiceName]struct {
 		Matchsets []struct {
